Refactoring for App/Gui separation
