+++ SheetMetalFormingCmd.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
 #
-#  SheetMetalCmd.py
+#  SheetMetalFormingCmd.py
 #  
 #  Copyright 2015 Shai Seger <shaise at gmail dot com>
 #  
@@ -23,186 +23,17 @@
 #  
 ###################################################################################
 
+import FreeCAD, Part, os, SheetMetalTools
 from FreeCAD import Gui
 from PySide import QtCore, QtGui
+from SheetMetalForming import SMBendWall
 
-import FreeCAD, FreeCADGui, Part, os, math
-import SheetMetalBaseCmd
-
-from SheetMetalLogger import SMLogger, UnfoldException, BendException, TreeException
-
-__dir__ = os.path.dirname(__file__)
-iconPath = os.path.join( __dir__, 'Resources', 'icons' )
-smEpsilon = 0.0000001
+icons_path = SheetMetalTools.icons_path
 
 # add translations path
-LanguagePath = os.path.join( __dir__, 'translations')
-Gui.addLanguagePath(LanguagePath)
+Gui.addLanguagePath(SheetMetalTools.language_path)
 Gui.updateLocale()
 
-def smWarnDialog(msg):
-    diag = QtGui.QMessageBox(
-        QtGui.QMessageBox.Warning,
-        FreeCAD.Qt.translate("QMessageBox", "Error in macro MessageBox"),
-        msg,
-    )
-    diag.setWindowModality(QtCore.Qt.ApplicationModal)
-    diag.exec_()
-
-def smBelongToBody(item, body):
-    if (body is None):
-        return False
-    for obj in body.Group:
-        if obj.Name == item.Name:
-            return True
-    return False
-
-def smIsPartDesign(obj):
-    return str(obj).find("<PartDesign::") == 0
-        
-def smIsOperationLegal(body, selobj):
-    #FreeCAD.Console.PrintLog(str(selobj) + " " + str(body) + " " + str(smBelongToBody(selobj, body)) + "\n")
-    if smIsPartDesign(selobj) and not smBelongToBody(selobj, body):
-        smWarnDialog(
-            FreeCAD.Qt.translate(
-                "QMessageBox",
-                "The selected geometry does not belong to the active Body.\n"
-                "Please make the container of this item active by\n"
-                "double clicking on it.",
-            )
-        )
-        return False
-    return True
- 
-def smthk(obj, foldface) :
-  normal = foldface.normalAt(0,0)
-  theVol = obj.Volume
-  if theVol < 0.0001:
-      SMLogger.error(
-            FreeCAD.Qt.translate(
-                "Logger",
-                "Shape is not a real 3D-object or to small for a metal-sheet!"
-            )
-        )
-  else:
-      # Make a first estimate of the thickness
-      estimated_thk = theVol/(obj.Area / 2.0)
-  #p1 = foldface.CenterOfMass
-  for v in foldface.Vertexes :
-    p1 = v.Point
-    p2 = p1 + estimated_thk * -1.5 * normal
-    e1 = Part.makeLine(p1, p2)
-    thkedge = obj.common(e1)
-    thk = thkedge.Length
-    if thk > smEpsilon :
-      break
-  return thk
-
-def angleBetween(ve1, ve2):
-  # Find angle between two vectors in degrees
-  return math.degrees(ve1.getAngle(ve2))
-
-def face_direction(face):
-  yL = face.CenterOfMass
-  uv = face.Surface.parameter(yL)
-  nv = face.normalAt(uv[0], uv[1])
-  direction = yL.sub(nv + yL)
-  #print([direction, yL])
-  return direction, yL
-
-def transform_tool(tool, base_face, tool_face, point = FreeCAD.Vector(0, 0, 0), angle = 0.0):
-  # Find normal of faces & center to align faces
-  direction1,yL1 = face_direction(base_face)
-  direction2,yL2 = face_direction(tool_face)
-
-  # Find angle between faces, axis of rotation & center of axis
-  rot_angle = angleBetween(direction1, direction2)
-  rot_axis = direction1.cross(direction2)
-  if rot_axis == FreeCAD.Vector (0.0, 0.0, 0.0):
-      rot_axis = FreeCAD.Vector(0, 1, 0).cross(direction2)
-  rot_center = yL2
-  #print([rot_center, rot_axis, rot_angle])
-  tool.rotate(rot_center, rot_axis, -rot_angle)
-  tool.translate(-yL2 + yL1)
-  #Part.show(tool, "tool")
-
-  tool.rotate(yL1, direction1, angle)
-  tool.translate(point)
-  #Part.show(tool,"tool")
-  return tool
-
-def makeforming(tool, base, base_face, thk, tool_faces = None, point = FreeCAD.Vector(0, 0, 0), angle = 0.0) :
-##  faces = [ face for face in tool.Shape.Faces for tool_face in tool_faces if not(face.isSame(tool_face)) ]
-#  faces = [ face for face in tool.Shape.Faces if not face in tool_faces ]
-#  tool_shell = Part.makeShell(faces)
-#  offsetshell = tool_shell.makeOffsetShape(thk, 0.0, inter = False, self_inter = False, offsetMode = 0, join = 2, fill = True)
-  offsetshell = tool.makeThickness(tool_faces, thk, 0.0001, False, False, 0, 0)
-  cutSolid = tool.fuse(offsetshell)
-  offsetshell_tran = transform_tool(offsetshell, base_face, tool_faces[0], point, angle)
-  #Part.show(offsetshell1, "offsetshell1")
-  cutSolid_trans = transform_tool(cutSolid, base_face, tool_faces[0], point, angle)
-  base = base.cut(cutSolid_trans)
-  base = base.fuse(offsetshell_tran)
-  #base.removeSplitter()
-  #Part.show(base, "base")
-  return base
-
-class SMBendWall:
-  def __init__(self, obj):
-    '''"Add Forming Wall" '''
-    selobj = Gui.Selection.getSelectionEx()
-
-    _tip_ = FreeCAD.Qt.translate("App::Property","Offset from Center of Face")
-    obj.addProperty("App::PropertyVectorDistance","offset","Parameters",_tip_)
-    _tip_ = FreeCAD.Qt.translate("App::Property","Suppress Forming Feature")
-    obj.addProperty("App::PropertyBool","SuppressFeature","Parameters",_tip_).SuppressFeature = False
-    _tip_ = FreeCAD.Qt.translate("App::Property","Tool Position Angle")
-    obj.addProperty("App::PropertyAngle","angle","Parameters",_tip_).angle = 0.0
-    _tip_ = FreeCAD.Qt.translate("App::Property","Thickness of Sheetmetal")
-    obj.addProperty("App::PropertyDistance","thickness","Parameters",_tip_)
-    _tip_ = FreeCAD.Qt.translate("App::Property","Base Object")
-    obj.addProperty("App::PropertyLinkSub", "baseObject", "Parameters",_tip_).baseObject = (selobj[0].Object, selobj[0].SubElementNames)
-    _tip_ = FreeCAD.Qt.translate("App::Property","Forming Tool Object")
-    obj.addProperty("App::PropertyLinkSub", "toolObject", "Parameters",_tip_).toolObject = (selobj[1].Object, selobj[1].SubElementNames)
-    _tip_ = FreeCAD.Qt.translate("App::Property","Point Sketch on Sheetmetal")
-    obj.addProperty("App::PropertyLink","Sketch","Parameters1",_tip_)
-    obj.Proxy = self
-
-  def execute(self, fp):
-    '''"Print a short message when doing a recomputation, this method is mandatory" '''
-
-    base = fp.baseObject[0].Shape
-    base_face = base.getElement(fp.baseObject[1][0])
-    thk = smthk(base, base_face)
-    fp.thickness = thk
-    tool = fp.toolObject[0].Shape
-    tool_faces = [tool.getElement(fp.toolObject[1][i]) for i in range(len(fp.toolObject[1]))]
-
-    offsetlist = []
-    if fp.Sketch:
-      sketch = fp.Sketch.Shape
-      for e in sketch.Edges:
-          #print(type(e.Curve))
-          if isinstance(e.Curve, (Part.Circle, Part.ArcOfCircle)):
-            pt1 = base_face.CenterOfMass
-            pt2 = e.Curve.Center
-            offsetPoint = pt2 - pt1
-            #print(offsetPoint)
-            offsetlist.append(offsetPoint)
-    else:
-      offsetlist.append(fp.offset)
-
-    if not(fp.SuppressFeature) :
-      for i in range(len(offsetlist)):
-        a = makeforming(tool, base, base_face, thk, tool_faces, offsetlist[i], fp.angle.Value)
-        base = a
-    else :
-      a = base
-    fp.Shape = a
-    Gui.ActiveDocument.getObject(fp.baseObject[0].Name).Visibility = False
-    Gui.ActiveDocument.getObject(fp.toolObject[0].Name).Visibility = False
-    if fp.Sketch:
-     Gui.ActiveDocument.getObject(fp.Sketch.Name).Visibility = False
 
 class SMFormingVP:
   "A View provider that nests children objects under the created one"
@@ -268,7 +99,7 @@ class SMFormingVP:
     return objs
 
   def getIcon(self):
-    return os.path.join( iconPath , 'SheetMetal_Forming.svg')
+    return os.path.join( icons_path , 'SheetMetal_Forming.svg')
 
   def setEdit(self,vobj,mode):
     taskd = SMFormingWallTaskPanel()
@@ -277,11 +108,11 @@ class SMFormingVP:
     self.Object.ViewObject.Visibility=False
     self.Object.baseObject[0].ViewObject.Visibility=True
     self.Object.toolObject[0].ViewObject.Visibility=True
-    FreeCADGui.Control.showDialog(taskd)
+    Gui.Control.showDialog(taskd)
     return True
 
   def unsetEdit(self,vobj,mode):
-    FreeCADGui.Control.closeDialog()
+    Gui.Control.closeDialog()
     self.Object.baseObject[0].ViewObject.Visibility=False
     self.Object.toolObject[0].ViewObject.Visibility=False
     self.Object.ViewObject.Visibility=True
@@ -337,7 +168,7 @@ class SMFormingPDVP:
     return objs
 
   def getIcon(self):
-    return os.path.join( iconPath , 'SheetMetal_Forming.svg')
+    return os.path.join( icons_path , 'SheetMetal_Forming.svg')
 
   def setEdit(self,vobj,mode):
     taskd = SMFormingWallTaskPanel()
@@ -346,11 +177,11 @@ class SMFormingPDVP:
     self.Object.ViewObject.Visibility=False
     self.Object.baseObject[0].ViewObject.Visibility=True
     self.Object.toolObject[0].ViewObject.Visibility=False
-    FreeCADGui.Control.showDialog(taskd)
+    Gui.Control.showDialog(taskd)
     return True
 
   def unsetEdit(self,vobj,mode):
-    FreeCADGui.Control.closeDialog()
+    Gui.Control.closeDialog()
     self.Object.baseObject[0].ViewObject.Visibility=False
     self.Object.toolObject[0].ViewObject.Visibility=False
     self.Object.ViewObject.Visibility=True
@@ -379,7 +210,7 @@ class SMFormingWallTaskPanel:
       # buttons
       self.addButton = QtGui.QPushButton(self.form)
       self.addButton.setObjectName("addButton")
-      self.addButton.setIcon(QtGui.QIcon(os.path.join( iconPath , 'SheetMetal_Update.svg')))
+      self.addButton.setIcon(QtGui.QIcon(os.path.join( icons_path , 'SheetMetal_Update.svg')))
       self.grid.addWidget(self.addButton, 3, 0, 1, 2)
 
       QtCore.QObject.connect(self.addButton, QtCore.SIGNAL("clicked()"), self.updateElement)
@@ -429,7 +260,7 @@ class SMFormingWallTaskPanel:
 
     def updateElement(self):
       if self.obj:
-        sel = FreeCADGui.Selection.getSelectionEx()[0]
+        sel = Gui.Selection.getSelectionEx()[0]
         if sel.HasSubObjects:
           obj = sel.Object
           for elt in sel.SubElementNames:
@@ -447,7 +278,7 @@ class SMFormingWallTaskPanel:
               if not found:
                 self.obj.baseObject = (sel.Object, sel.SubElementNames)
 
-        sel = FreeCADGui.Selection.getSelectionEx()[1]
+        sel = Gui.Selection.getSelectionEx()[1]
         if sel.HasSubObjects:
           obj = sel.Object
           for elt in sel.SubElementNames:
@@ -468,7 +299,7 @@ class SMFormingWallTaskPanel:
 
     def accept(self):
       FreeCAD.ActiveDocument.recompute()
-      FreeCADGui.ActiveDocument.resetEdit()
+      Gui.ActiveDocument.resetEdit()
       #self.obj.ViewObject.Visibility=True
       return True
 
@@ -481,7 +312,7 @@ class AddFormingWallCommand():
   """Add Forming Wall command"""
 
   def GetResources(self):
-    return {'Pixmap'  : os.path.join( iconPath , 'SheetMetal_Forming.svg') , # the name of a svg file available in the resources
+    return {'Pixmap'  : os.path.join( icons_path , 'SheetMetal_Forming.svg') , # the name of a svg file available in the resources
             'MenuText': FreeCAD.Qt.translate('SheetMetal','Make Forming in Wall') ,
             'Accel': "M, F",
             'ToolTip' : FreeCAD.Qt.translate('SheetMetal','Make a forming using tool in metal sheet\n'
@@ -494,24 +325,29 @@ class AddFormingWallCommand():
     doc = FreeCAD.ActiveDocument
     view = Gui.ActiveDocument.ActiveView
     activeBody = None
+    sel = Gui.Selection.getSelectionEx()
     selobj = Gui.Selection.getSelectionEx()[0].Object
-    viewConf = SheetMetalBaseCmd.GetViewConfig(selobj)
+    viewConf = SheetMetalTools.GetViewConfig(selobj)
     if hasattr(view,'getActiveObject'):
       activeBody = view.getActiveObject('pdbody')
-    if not smIsOperationLegal(activeBody, selobj):
+    if not SheetMetalTools.smIsOperationLegal(activeBody, selobj):
         return
     doc.openTransaction("WallForming")
-    if activeBody is None or not smIsPartDesign(selobj):
+    if activeBody is None or not SheetMetalTools.smIsPartDesign(selobj):
       a = doc.addObject("Part::FeaturePython","WallForming")
       SMBendWall(a)
+      a.baseObject = (selobj, sel[0].SubElementNames)
+      a.toolObject = (sel[1].Object, sel[1].SubElementNames)
       SMFormingVP(a.ViewObject)
     else:
       #FreeCAD.Console.PrintLog("found active body: " + activeBody.Name)
       a = doc.addObject("PartDesign::FeaturePython","WallForming")
       SMBendWall(a)
+      a.baseObject = (selobj, sel[0].SubElementNames)
+      a.toolObject = (sel[1].Object, sel[1].SubElementNames)
       SMFormingPDVP(a.ViewObject)
       activeBody.addObject(a)
-    SheetMetalBaseCmd.SetViewConfig(a, viewConf)
+    SheetMetalTools.SetViewConfig(a, viewConf)
     doc.recompute()
     doc.commitTransaction()
     return
