+++ SheetMetalJunctionCmd.py
@@ -1,22 +1,22 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
 #
-#  SheetMetalExtendCmd.py
+#  SheetMetalJunction.py
 #
-#  Copyright 2020 Jaise James <jaisekjames at gmail dot com>
+#  Copyright 2015 Shai Seger <shaise at gmail dot com>
 #
-#  This program is free software; you can redistribute it and/or
-#  modify it under the terms of the GNU Lesser General Public
-#  License as published by the Free Software Foundation; either
-#  version 2 of the License, or (at your option) any later version.
+#  This program is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation; either version 2 of the License, or
+#  (at your option) any later version.
 #
 #  This program is distributed in the hope that it will be useful,
 #  but WITHOUT ANY WARRANTY; without even the implied warranty of
 #  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #  GNU General Public License for more details.
 #
-#  You should have received a copy of the GNU Lesser General Public
-#  License along with this program; if not, write to the Free Software
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software
 #  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
 #  MA 02110-1301, USA.
 #
@@ -26,11 +26,16 @@
 import FreeCAD, Part, os, SheetMetalTools
 from FreeCAD import Gui
 from PySide import QtCore, QtGui
-from SheetMetalExtend import SMExtrudeWall
+from SheetMetalJunction import SMJunction
 
 icons_path = SheetMetalTools.icons_path
 
-class SMViewProviderTree:
+# add translations path
+Gui.addLanguagePath(SheetMetalTools.language_path)
+Gui.updateLocale()
+
+
+class SMJViewProviderTree:
   "A View provider that nests children objects under the created one"
 
   def __init__(self, obj):
@@ -87,15 +92,13 @@ class SMViewProviderTree:
     objs = []
     if hasattr(self.Object,"baseObject"):
       objs.append(self.Object.baseObject[0])
-    if hasattr(self.Object,"Sketch"):
-      objs.append(self.Object.Sketch)
     return objs
 
   def getIcon(self):
-    return os.path.join( icons_path , 'SheetMetal_Extrude.svg')
+    return os.path.join( icons_path , 'SheetMetal_AddJunction.svg')
 
   def setEdit(self,vobj,mode):
-    taskd = SMBendWallTaskPanel()
+    taskd = SMJunctionTaskPanel()
     taskd.obj = vobj.Object
     taskd.update()
     self.Object.ViewObject.Visibility=False
@@ -109,8 +112,8 @@ class SMViewProviderTree:
     self.Object.ViewObject.Visibility=True
     return False
 
-class SMViewProviderFlat:
-  "A View provider that places objects flat under base object"
+class SMJViewProviderFlat:
+  "A View provider that nests children objects under the created one"
 
   def __init__(self, obj):
     obj.Proxy = self
@@ -151,16 +154,13 @@ class SMViewProviderFlat:
       self.Object = doc.getObject(state['ObjectName'])
 
   def claimChildren(self):
-    objs = []
-    if hasattr(self.Object,"Sketch"):
-      objs.append(self.Object.Sketch)
-    return objs
+    return []
 
   def getIcon(self):
-    return os.path.join( icons_path , 'SheetMetal_Extrude.svg')
+    return os.path.join( icons_path , 'SheetMetal_AddJunction.svg')
 
   def setEdit(self,vobj,mode):
-    taskd = SMBendWallTaskPanel()
+    taskd = SMJunctionTaskPanel()
     taskd.obj = vobj.Object
     taskd.update()
     self.Object.ViewObject.Visibility=False
@@ -174,19 +174,19 @@ class SMViewProviderFlat:
     self.Object.ViewObject.Visibility=True
     return False
 
-class SMBendWallTaskPanel:
+class SMJunctionTaskPanel:
     '''A TaskPanel for the Sheetmetal'''
     def __init__(self):
 
       self.obj = None
       self.form = QtGui.QWidget()
-      self.form.setObjectName("SMBendWallTaskPanel")
-      self.form.setWindowTitle("Binded faces/edges list")
+      self.form.setObjectName("SMJunctionTaskPanel")
+      self.form.setWindowTitle("Binded edges list")
       self.grid = QtGui.QGridLayout(self.form)
       self.grid.setObjectName("grid")
       self.title = QtGui.QLabel(self.form)
       self.grid.addWidget(self.title, 0, 0, 1, 2)
-      self.title.setText("Select new face(s)/Edge(s) and press Update")
+      self.title.setText("Select new Edge(s) and press Update")
 
       # tree
       self.tree = QtGui.QTreeWidget(self.form)
@@ -237,16 +237,16 @@ class SMBendWallTaskPanel:
         if sel.HasSubObjects:
           obj = sel.Object
           for elt in sel.SubElementNames:
-            if "Face" in elt or "Edge" in elt:
-              face = self.obj.baseObject
+            if "Edge" in elt:
+              edge = self.obj.baseObject
               found = False
-              if (face[0] == obj.Name):
-                if isinstance(face[1],tuple):
-                  for subf in face[1]:
+              if (edge[0] == obj.Name):
+                if isinstance(edge[1],tuple):
+                  for subf in edge[1]:
                     if subf == elt:
                       found = True
                 else:
-                  if (face[1][0] == elt):
+                  if (edge[1][0] == elt):
                     found = True
               if not found:
                 self.obj.baseObject = (sel.Object, sel.SubElementNames)
@@ -259,45 +259,46 @@ class SMBendWallTaskPanel:
         return True
 
     def retranslateUi(self, TaskPanel):
-        #TaskPanel.setWindowTitle(QtGui.QApplication.translate("draft", "Faces", None))
+        #TaskPanel.setWindowTitle(QtGui.QApplication.translate("draft", "edges", None))
         self.addButton.setText(QtGui.QApplication.translate("draft", "Update", None))
 
-class SMExtrudeCommandClass():
-  """Extrude face"""
+
+class AddJunctionCommandClass():
+  """Add Junction command"""
 
   def GetResources(self):
-    return {'Pixmap'  : os.path.join( icons_path , 'SheetMetal_Extrude.svg'), # the name of a svg file available in the resources
-            'MenuText': FreeCAD.Qt.translate('SheetMetal','Extend Face'),
-            'Accel': "E",
-            'ToolTip' : FreeCAD.Qt.translate('SheetMetal','Extends one or more face, on existing sheet metal.\n'
-            '1. Select edges or thickness side faces to create walls.\n'
-            '2. Select a sketch in property editor to create tabs. \n'
-            '3. Use Property editor to modify other parameters')}
+    return {'Pixmap'  : os.path.join( icons_path , 'SheetMetal_AddJunction.svg'), # the name of a svg file available in the resources
+            'MenuText': FreeCAD.Qt.translate('SheetMetal','Make Junction'),
+            'Accel': "S, J",
+            'ToolTip' : FreeCAD.Qt.translate('SheetMetal','Create a rip where two walls come together on solids.\n'
+            '1. Select edge(s) to create rip on corner edge(s).\n'
+            '2. Use Property editor to modify parameters')}
 
   def Activated(self):
     doc = FreeCAD.ActiveDocument
     view = Gui.ActiveDocument.ActiveView
     activeBody = None
     sel = Gui.Selection.getSelectionEx()[0]
-    selobj = sel.Object
+    selobj = Gui.Selection.getSelectionEx()[0].Object
     viewConf = SheetMetalTools.GetViewConfig(selobj)
     if hasattr(view,'getActiveObject'):
       activeBody = view.getActiveObject('pdbody')
     if not SheetMetalTools.smIsOperationLegal(activeBody, selobj):
         return
-    doc.openTransaction("Extend")
-    if (activeBody is None):
-      a = doc.addObject("Part::FeaturePython","Extend")
-      SMExtrudeWall(a)
+    doc.openTransaction("Add Junction")
+    if activeBody is None or not SheetMetalTools.smIsPartDesign(selobj):
+      a = doc.addObject("Part::FeaturePython","Junction")
+      SMJunction(a)
       a.baseObject = (selobj, sel.SubElementNames)
-      SMViewProviderTree(a.ViewObject)
+      SMJViewProviderTree(a.ViewObject)
     else:
-      a = doc.addObject("PartDesign::FeaturePython","Extend")
-      SMExtrudeWall(a)
+      #FreeCAD.Console.PrintLog("found active body: " + activeBody.Name)
+      a = doc.addObject("PartDesign::FeaturePython","Junction")
+      SMJunction(a)
       a.baseObject = (selobj, sel.SubElementNames)
-      SMViewProviderFlat(a.ViewObject)
+      SMJViewProviderFlat(a.ViewObject)
       activeBody.addObject(a)
-    SheetMetalTools.SetViewConfig(a, viewConf)    
+    SheetMetalTools.SetViewConfig(a, viewConf)
     Gui.Selection.clearSelection()
     doc.recompute()
     doc.commitTransaction()
@@ -306,12 +307,11 @@ class SMExtrudeCommandClass():
   def IsActive(self):
     if len(Gui.Selection.getSelection()) < 1 or len(Gui.Selection.getSelectionEx()[0].SubElementNames) < 1:
       return False
-    selobj = Gui.Selection.getSelection()[0]
-    if selobj.isDerivedFrom("Sketcher::SketchObject"):
-      return False
-    for selFace in Gui.Selection.getSelectionEx()[0].SubObjects:
-      if type(selFace) == Part.Vertex :
+#    selobj = Gui.Selection.getSelection()[0]
+    for selEdge in Gui.Selection.getSelectionEx()[0].SubObjects:
+      if type(selEdge) != Part.Edge :
         return False
     return True
 
-Gui.addCommand("SheetMetal_Extrude", SMExtrudeCommandClass())
+Gui.addCommand("SheetMetal_AddJunction", AddJunctionCommandClass())
+
